+++ mono-mm/mono-mm/Spectrum/SpectrumFont.cs
@@ -14,7 +14,7 @@ namespace MonoManicMiner.Spectrum
 
         }
 
-        public override void Draw(SpriteBatch spriteBatch, float scale)
+        public override void Draw(SpriteBatch spriteBatch, int scale)
         {
             int colourOffset = 0;
